Hexen: add support for extended nodes
Fixes "li->frontsector" issue.

+++ src/hexen/r_local.h
@@ -161,14 +161,15 @@ typedef struct line_s
     vertex_t *v2;
     fixed_t dx;
     fixed_t dy;
-    short flags;
+    unsigned short flags; // [crispy] extended nodes
     byte special;
     byte arg1;
     byte arg2;
     byte arg3;
     byte arg4;
     byte arg5;
-    short sidenum[2];
+    // sidenum[1] will be NO_INDEX if one sided
+    unsigned short sidenum[2]; // [crispy] extended nodes
     fixed_t bbox[4];
     slopetype_t slopetype;
     sector_t *frontsector;
@@ -234,8 +235,8 @@ typedef struct polyblock_s
 typedef struct subsector_s
 {
     sector_t *sector;
-    short numlines;
-    short firstline;
+    int numlines; // [crispy] extended nodes
+    int firstline; // [crispy] extended nodes
     polyobj_t *poly;
 } subsector_t;
 
@@ -243,7 +244,8 @@ typedef struct
 {
     fixed_t x, y, dx, dy;       // partition line
     fixed_t bbox[2][4];         // bounding box for each child
-    unsigned short children[2]; // if NF_SUBSECTOR its a subsector
+    // if NF_SUBSECTOR its a subsector
+    int children[2]; // [crispy] extended nodes
 } node_t;
